fix(routers): address review — VLM thinking forwarding + shared splitter
- _apply_chat_template_vlm now forwards enable_thinking on the no-tools VLM
  path (mlx_vlm.apply_chat_template forwards **kwargs to the tokenizer), so
  the toggle is no longer silently dropped for VLM models.
- Extract the streaming thinking-splitter into routers/thinking_split.py,
  shared explicitly by the chat and generate routers instead of importing
  private chat.py helpers.
